Try silhouette analysis on all dimensions
Results inconclusive to generate good clusters. Might be because there are 3400 dimensions for 18k records, but we don't have a statistical "high dimensional" problem since number of dimensions p is still less than number of records n.

Kept trying on Render, but memory limit is very real and hard to get over. Migrate to something else (Digital Ocean)

To do:

1. Get ready to migrate MeaLeon
2. Do some form of dimension reduction
3. Prep for Optuna for hyperparameter tuning
4. Prep for some form of MLOps (MLFlow or Weights and Biases)