Update environment with FlagEmbedding
Forced restart cleared jupyter kernel, so might as well add FlagEmbedding library to create embeddings for retrieval.

Read a few discussions on how slow Stanza is with the pipeline. One thing is that I think there is a strong benefit to using lemmas for tokens, which needs many of the processing pipelines. Unfortunately, this looks like the slow part of processing.

Saw this discussion on [using multiprocessing combined with Stanza](https://github.com/stanfordnlp/stanza/issues/552) and it seems like multiprocessing with GPU doesn't result in much performance increase but does drastically cause the GPU to work harder. However, discussion reminded me that I turned off GPU with Stanza in an attempt to use BERTopic with Stanza. Will re-enable and retime.