Complete first video in playlist
We coded along the [Installation/API Routes/Swagger Docs video](https://youtu.be/Lw-zLopB3o0?si=I6AaS1jCxhnR-QGb), did notice that this is basically a video form of the [Quick Start tutorial at FastAPI](https://fastapi.tiangolo.com), but we added comments in the `main.py` file to help be aware of how to activate the app.

We also updated the Poetry environment to get FastAPI working. Needed uvicorn as well